Copy WebBrowserExtensions to InternalWebBrowserExtensions Create empty WebBrowserExtensions to house common methods
